HomeWork2_task_three_changed

+++ src/ru/skypro/Main.java
@@ -6,7 +6,7 @@ public class Main {
 
     public static void main(String[] args) {
 short bananas = 5;
-short milk = 200;
+float milk = 200F;
 short plombir = 2;
 short eggs = 4;
 float finalWeight = bananas*80 + milk/100*105 + plombir*100 + eggs*70;
